backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10
flask-socketio==5.3.6
python-dotenv==1.0.0
werkzeug==2.2.3
python-engineio==4.2.1
python-socketio==5.4.0
python-chess==1.999
openai==1.12.0
anthropic==0.18.0
google-generativeai==0.3.2
katago==1.0.0
